修复interval数据溢出问题

+++ src/common/backend/utils/adt/datetime.cpp
@@ -2799,7 +2799,7 @@ int DecodeInterval(char** field, const int* ftype, int nf, int range, int* dtype
                     type = DTK_MONTH;
                     if (*field[i] == '-')
                         val2 = -val2;
-                    long int monthes = val * MONTHS_PER_YEAR + val2;
+                    long int monthes = (long)val * MONTHS_PER_YEAR + val2;
                     if (monthes > PG_INT32_MAX || monthes < PG_INT32_MIN) {
                         return DTERR_FIELD_OVERFLOW;
                     }
